+++ src/main/java/com/qinyuan15/utils/mvc/controller/PaginationItemFactory.java
@@ -7,7 +7,7 @@ import java.util.List;
  * Created by qinyuan on 15-4-5.
  */
 public interface PaginationItemFactory<T> {
-    long getCount();
+    int getCount();
 
     List<T> getInstances(int firstResult, int maxResults);
 }
